Split avatar endpoints into multiple files
