+++ server/main.js
@@ -8,7 +8,6 @@ import './lib/cordova';
 import './lib/roomFiles';
 import './startup/migrations';
 import './startup/appcache';
-import './startup/avatar';
 import './startup/cron';
 import './startup/initialData';
 import './startup/presence';
@@ -78,5 +77,6 @@ import './publications/subscription';
 import './publications/userAutocomplete';
 import './publications/userChannels';
 import './publications/userData';
+import './routes/avatar';
 import './stream/messages';
 import './stream/streamBroadcast';
