+++ server/routes/avatar/index.js
@@ -1,0 +1,4 @@
+import './middlewares';
+
+import './room';
+import './user';
