+++ server/routes/avatar/middlewares/index.js
@@ -1,0 +1,1 @@
+import './auth';
